+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Poll files node: Added `pollIntervalSeconds` field to output payload alongside `nextPoll`
+-   Poll files node: "All results" output is now hidden by default, with a checkbox option to enable it
 
 ## [1.4.1] - 2025-11-16